refactor: Move LevelProgress to shared kernel
